DocTour/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
imageio==2.34.2
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
networkx==3.3
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pip==24.0
scikit-image==0.24.0
scipy==1.14.0
tifffile==2024.7.24
Werkzeug==3.0.3
